# docs/contracts/ai_service_api_contract.yaml
openapi: 3.0.3
info:
  title: AI Service API
  description: |
    Görsel analiz, vektör benzerlik araması, agent tabanlı sohbet ve güvenlik kontrolleri
    sağlayan servis için API sözleşmesi.
  version: 1.0.0

servers:
  - url: https://api.example.com
    description: Production

security:
  - ApiKeyAuth: []

paths:
  /analyze:
    post:
      summary: Görsel Analiz (Async Job)
      description: |
        Görsel analiz işlemini asenkron bir job olarak başlatır.
        İş Redis queue'ya alınır ve hemen bir job_id döndürülür.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '202':
          description: Analiz job'u kuyruğa alındı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobResponse'
        '400':
          description: Geçersiz istek
        '500':
          description: Sunucu hatası

  /analyze/{job_id}:
    get:
      summary: Görsel Analiz Job Durumu
      description: |
        Belirtilen job_id için analiz durumunu ve varsa sonucu döndürür.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job durumu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobStatusResponse'
        '404':
          description: Job bulunamadı
        '500':
          description: Sunucu hatası

  /upsert:
    post:
      summary: Vektör Upsert
      description: |
        Yeni bir post görseli için vektör ve metadata kaydı oluşturur veya günceller.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRequest'
      responses:
        '200':
          description: Başarılı kayıt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpsertResponse'
        '500':
          description: Sunucu hatası

  /delete-record:
    post:
      summary: Vektör Silme
      description: |
        Verilen post_id'ye ait vektör kaydını siler.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
      responses:
        '200':
          description: Başarılı silme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '500':
          description: Sunucu hatası

  /sim-search:
    post:
      summary: Metin Tabanlı Benzerlik Araması
      description: |
        Metin sorgusunu vektöre çevirir ve vector database
        üzerinde en benzer sonuçları döndürür.
        En az bir tanesi (query_text veya image_url) sağlanmalıdır.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimSearchRequest'
      responses:
        '200':
          description: Benzerlik araması sonuçları
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimSearchResponse'
        '400':
          description: Geçersiz istek
        '500':
          description: Sunucu hatası

  /chat:
    post:
      summary: Agent Sohbet
      description: |
        Kullanıcıyla bağlam farkındalığı olan sohbet sağlar.
        Gerekirse ek aksiyonlar tetikleyebilir.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Sohbet cevabı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '500':
          description: Sunucu hatası

  /nsfw-check:
    post:
      summary: NSFW İçerik Kontrolü
      description: |
        Görsel üzerinde NSFW içerik tespiti yapar ve güven skoru döndürür.
        Skor 0.0 (güvenli) ile 1.0 (NSFW) arasında bir değerdir.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NSFWCheckRequest'
      responses:
        '200':
          description: NSFW kontrolü sonucu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NSFWCheckResponse'
        '500':
          description: Sunucu hatası

  /health:
    get:
      summary: Sistem Durumu
      description: Servisin ve modellerin ayakta olup olmadığını kontrol eder.
      security: []
      responses:
        '200':
          description: Sistem durumu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

components:
  schemas:
    AnalyzeRequest:
      type: object
      required:
        - post_id
        - image_url
      properties:
        post_id:
          type: string
          description: Analiz edilecek postun ID'si
        image_url:
          type: string
          format: uri
          description: Görselin URL adresi

    UpsertItem:
      type: object
      required:
        - post_id
        - image_url
      properties:
        post_id:
          type: string
        image_url:
          type: string
          format: uri

    UpsertRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UpsertItem'

    UpsertResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, failed]
        count:
          type: integer

    DeleteRequest:
      type: object
      required:
        - post_id
      properties:
        post_id:
          type: string

    DeleteResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, failed]
        vector_id:
          type: string

    AnalyzeJobResponse:
      type: object
      properties:
        job_id:
          type: string
          description: Oluşturulan job ID
        status:
          type: string
          example: queued

    AnalyzeJobStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [queued, processing, completed, failed]
        result:
          type: object
          nullable: true
          properties:
            aesthetic_score:
              type: number
              format: float
            suggested_tags:
              type: array
              items:
                type: string
        error:
          type: string
          nullable: true

    SimSearchRequest:
      type: object
      oneOf:
        - required: [query_text]
        - required: [image_url]
      properties:
        query_text:
          type: string
          description: Text query for similarity search (at least one of query_text or image_url must be provided)
        image_url:
          type: string
          format: uri
          description: Image URL for similarity search (at least one of query_text or image_url must be provided)
        top_k:
          type: integer
          minimum: 1
          default: 5
          description: Number of results to return

    SimSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              image_url:
                type: string
                format: uri
              sim_score:
                type: number
                format: float

    ChatRequest:
      type: object
      required:
        - user_id
        - message
      properties:
        user_id:
          type: string
        message:
          type: string
        history:
          type: array
          items:
            type: object
            description: Önceki mesajlar (serbest format)

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        actions:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Action'
    
    Action:
      type: object
      required:
        - type
        - parameters
      properties:
        type:
          type: string
          description: Action türü (örn. "search_images", "analyze_photo")
          example: search_images
        parameters:
          type: object
          description: Action için gerekli parametreler
          example:
            query: "sunset beach"
            top_k: 10

    NSFWCheckRequest:
      type: object
      required:
        - image_url
      properties:
        image_url:
          type: string
          format: uri
          description: Kontrol edilecek görselin URL'si

    NSFWCheckResponse:
      type: object
      properties:
        conf_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: NSFW güven skoru (0.0 = güvenli, 1.0 = NSFW)
          example: 0.15

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: online
        models_loaded:
          type: array
          items:
            type: string

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Required for all endpoints except /health.